gauge numbers are now the river kilometres

+++ src/flood_wave_handler.py
@@ -414,7 +414,7 @@ class FloodWaveHandler:
 
         for node in joined_graph.nodes():
             x_coord = (datetime.strptime(node[1], '%Y-%m-%d') - start).days - 1
-            y_coord = len(gauges) - gauges.index(int(node[0]))
+            y_coord = len(gauges) - gauges.index(float(node[0]))
             positions[node] = (x_coord, y_coord)
         return positions
